+++ src/plugins/curator/components/timeline-picker-field/constants.js
@@ -1,4 +1,3 @@
 export default {
   CHANNEL_NAME: 'timelinePicker',
-  REQUEST_COUNT: 10,
 };