+++ src/plugins/curator/components/widget-picker-field/constants.js
@@ -1,4 +1,3 @@
 export default {
   CHANNEL_NAME: 'widgetPicker',
-  REQUEST_COUNT: 5,
 };