+++ src/plugins/people/components/people-picker-field/constants.js
@@ -1,4 +1,3 @@
 export default {
   CHANNEL_NAME: 'peoplePicker',
-  REQUEST_COUNT: 10,
 };