+++ src/plugins/sys/components/redirect-picker-field/constants.js
@@ -1,4 +1,3 @@
 export default {
   CHANNEL_NAME: 'redirectPicker',
-  REQUEST_COUNT: 10,
 };